+++ selfdrive/modeld/runners/thneedmodel.h
@@ -5,7 +5,7 @@
 
 class ThneedModel : public RunModel {
 public:
-  ThneedModel(const char *path, float *loutput, size_t loutput_size, int runtime, bool luse_extra = false);
+  ThneedModel(const char *path, float *loutput, size_t loutput_size, int runtime, bool luse_extra = false, bool use_tf8 = false, cl_context context = NULL);
   void addRecurrent(float *state, int state_size);
   void addTrafficConvention(float *state, int state_size);
   void addDesire(float *state, int state_size);
@@ -14,8 +14,8 @@ public:
   void execute();
   void* getInputBuf();
   void* getExtraBuf();
-private:
   Thneed *thneed = NULL;
+private:
   bool recorded;
   bool use_extra;
