+++ selfdrive/modeld/thneed/thneed.h
@@ -16,6 +16,8 @@
 
 using namespace std;
 
+cl_int thneed_clSetKernelArg(cl_kernel kernel, cl_uint arg_index, size_t arg_size, const void *arg_value);
+
 namespace json11 {
   class Json;
 }
@@ -89,7 +91,7 @@ class CachedCommand: public CachedIoctl {
 
 class Thneed {
   public:
-    Thneed(bool do_clinit=false);
+    Thneed(bool do_clinit=false, cl_context _context = NULL);
     void stop();
     void execute(float **finputs, float *foutput, bool slow=false);
     void wait();
@@ -110,9 +112,12 @@ class Thneed {
     bool record = false;
     int debug;
     int timestamp;
+
+#ifdef QCOM2
     unique_ptr<GPUMalloc> ram;
     vector<unique_ptr<CachedIoctl> > cmds;
     int fd;
+#endif
 
     // all CL kernels
     void find_inputs_outputs();
